finished refactor - implementing cpp complete

+++ PSO/utils_.h
@@ -22,7 +22,7 @@ double get_normal(double mean, double var) {
 
 double compute_distance(std::vector<double> p_1,std::vector<double> p_2) {
     double sum=0;
-    for (int i=0; i<p_1.size(); i++) {
+    for (size_t i=0; i<p_1.size(); i++) {
         sum += std::pow((p_1[i]-p_2[i]),2);
     }
     return std::sqrt(sum);
@@ -33,10 +33,10 @@ double get_rand(double start, double end) {
         return distribution(generator);
 }
 
-int get_rand_idx(int start, int end){
+size_t get_rand_idx(int start, int end){
     // std::random_device rnd_device;
     std::default_random_engine generator(std::rand());
-    std::uniform_int_distribution<int> distribution(start, end);
+    std::uniform_int_distribution<size_t> distribution(start, end);
     int i = distribution(generator);
     // std::cout<<i<<"  iiii "<<start<<" "<<end<<std::endl;
     return distribution(generator);
